requirements.txt
django
django-isbn-field